==> 05_fine_tuning.ipynb <==
import torch
from peft import PeftModel
from transformers import AutoTokenizer, AutoModelForCausalLM, BitsAndBytesConfig, pipeline
import dotenv
import os 
import accelerate

import pandas as pd
import json
import pyarrow as pa
import pyarrow.dataset as ds
from datasets import Dataset


model_name = "serlikopar/Enlighten_Instruct"

dotenv.load_dotenv("hugging_face.env")

secret_hf = os.getenv('HUGGINGFACE_TOKEN')
!huggingface-cli login --token $secret_hf

tokenizer = AutoTokenizer.from_pretrained(base_model, trust_remote_code=True)
tokenizer.padding_side = 'right'
tokenizer.pad_token = tokenizer.eos_token
tokenizer.add_eos_token = True
tokenizer.bos_token, tokenizer.eos_token

tokenizer = AutoTokenizer.from_pretrained(model_name)
model = AutoModelForSequenceClassification.from_pretrained(model_name)

base_model = "mistralai/Mistral-7B-Instruct-v0.2"
new_model = "serlikopar/Enlighten_Instruct"
base_model_reload = AutoModelForCausalLM.from_pretrained(
        base_model,
        torch_dtype=torch.bfloat16,
        return_dict=True,
        low_cpu_mem_usage=True,
        device_map="auto",
        trust_remote_code=True,
)
model = PeftModel.from_pretrained(base_model_reload, new_model)
pipe = pipeline(task="text-generation", model=model, tokenizer=tokenizer, max_length=200)

import pandas as pd
import json

def make_df_from_jsonl(file_path):
    questions = []
    answers = []

    with open(file_path, 'r', encoding='utf-8') as file:
        for line_number, line in enumerate(file, start=1):
            try:
                json_data = json.loads(line)
                # Extract messages
                messages = json_data['messages']
                # Initialize placeholders for user and assistant messages
                question = next((msg['content'] for msg in messages if msg['role'] == 'user'), None)
                answer = next((msg['content'] for msg in messages if msg['role'] == 'assistant'), None)
                
                # Only append if both question and answer are found
                if question and answer:
                    questions.append(question.replace('\\n', '\n'))
                    answers.append(answer.replace('\\n', '\n'))
                
            except json.JSONDecodeError as e:
                print(f"Error parsing JSON on line {line_number}: {e}")
                continue

    # Create DataFrame with specific columns
    df = pd.DataFrame({'Question': questions, 'Answer': answers})
    
    return df




file_path = 'test_samples.jsonl'
df = make_df_from_jsonl(file_path)

df

df.head()

df.columns

df.to_csv("test_data_62_questions_with_formatting.csv", index=False)

df['text'] = '<s>[INST]' + df['Question'] +'[/INST]'+ df['Answer'] + '</s>'



df

df2 = df["text"]

df2 = pd.DataFrame(df2)

type(df2)

df2

df2.to_csv("test_data_62_questions_one_col.csv", index=False)
